+++ packages/sodium/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.1] - 2021-05-11
 ### Added
 - New libsodium API: crypto_auth
+- Add `Sodium.secureCopy` (#1)
 
 ## [0.1.0] - 2021-05-09
 ### Added